Patches to enable Hedge in Linux v5.5
Linux v5.5 and older does not implement the kvm_vcpu_destory function,
hence we need to use the kvm_arch_vcpu_destroy function. This function
is declared in the KVM header files, but is not exported.

The specific version v5.5 is used from raspberry pi kernels. However,
recently, they switched to v5.10.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>